clone selected repository.
