+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-samples/src/main/resources/META-INF/azure-intellij-plugin-samples-git4idea.xml
@@ -1,0 +1,8 @@
+<idea-plugin>
+    <extensions defaultExtensionNs="com.intellij">
+        <openapi.vcs.ui.cloneDialog.VcsCloneDialogExtension
+                id="com.microsoft.azure.toolkit.intellij.samples.view.AzureSamplesCloneDialogExtension"
+                implementation="com.microsoft.azure.toolkit.intellij.samples.view.AzureSamplesCloneDialogExtension"
+                order="after com.intellij.util.ui.cloneDialog.RepositoryUrlCloneDialogExtension"/>
+    </extensions>
+</idea-plugin>
